Read ChatbotTrainer.cpp end to end (second pass)
Full re-read of src/ChatbotTrainer.cpp (1998 lines) end to end as part of
a full independent re-verification of the whole prior audit. The core
training loop (train_epoch(), validate(), the LR schedules, TD-013's
Welford/outlier/weight-update-ratio diagnostics, TD-017's adaptive
gradient clipping, TD-023's async BLEU/ROUGE scoring, the tokenized-data
cache, early stopping/best-weight restore) all checked out correct,
including several already-documented and fixed prior bugs (TD-062's
update-count drift, TD-023's clone()-based async scoring) whose fixes and
rationale comments were independently re-verified as still consistent
with the surrounding code.

Found and fixed one small logging-only bug: validate_and_correct_config()
truncated config.min_learning_rate (a small float, typically << 1, e.g.
0.001) to int before logging it in the "min_learning_rate >= learning_rate"
auto-correction warning — collapsing any realistic value to a misleading
0 in the log message. The actual correction itself (setting
min_learning_rate to 1% of the base learning rate) was always computed
correctly from the real float value; only the diagnostic message was
wrong. Fixed by keeping the pre-correction value as a float.

Verification: confirmed via a standalone repro (0.001f truncated to 0)
before fixing; rebuilt chatbottrainerTests and confirmed the full
ChatbotTrainerTests suite still passes (no test previously exercised this
log message's content, since it's a diagnostic-only fix with no behavior
change).

Bumped @adai-reviewed.

+++ src/ChatbotTrainer.cpp
@@ -1,6 +1,6 @@
 // @adai-status: beta        (capped by TD-039 — large, actively evolving core trainer)
 // @adai-version: 0.9.0
-// @adai-reviewed: 2026-09-08
+// @adai-reviewed: 2026-09-09
 
 #include "ChatbotTrainer.hpp"
 #include <algorithm>
@@ -268,7 +268,11 @@ void ChatbotTrainer::validate_and_correct_config() {
 
     // Validate min_learning_rate < learning_rate
     if (config.min_learning_rate >= config.learning_rate) {
-        int original_min_lr = static_cast<int>(config.min_learning_rate);
+        // min_learning_rate is a small float (typically << 1, e.g. 0.001) —
+        // truncating it to int before logging always collapses it to 0,
+        // making the warning below misleadingly show "min_learning_rate (0)"
+        // regardless of the actual prior value.
+        float original_min_lr = config.min_learning_rate;
         config.min_learning_rate = config.learning_rate * 0.01f;  // 1% of base LR
         adai::Logger::warn("⚠️  min_learning_rate ({}) >= learning_rate ({})", original_min_lr,
                            config.learning_rate);
